Move all "check_" files into "checks" subdirectory
This is good because is separates all of our 'checking' logic from
the scaffolding we are using to explore the PDF and create internal
data structures etc.
