+++ scanner/checks/alt_text.py
