+++ scanner/checks/figures.py
