+++ scanner/checks/tables.py
