+++ tests/test_check_title.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 from pikepdf import Pdf
 
-from scanner.checks import check_metadata_and_title
+from scanner.checks.document import check_metadata_and_title
 
 
 FIXTURE_SUBDIR = "title"
